requirement.txt.txt
beautifulsoup4==4.10.0
mysql-connector-python==8.0.28
gTTS==2.2.3
langdetect==1.0.9
translate==3.6.1
